+++ .gitlab-ci.yml
@@ -11,6 +11,6 @@ validate:
     - apk add --update python3 python3-dev py-pip build-base openjdk8-jre && pip install virtualenv && rm -rf /var/cache/apk/*
     - pip3 install -r requirements.txt
   script:
-    - python3 bin/test_ssa.py --skip-error tests/*/*
-    - python3 bin/validate_ssa_detections.py --skip-error detections/*/*
+    - python3 bin/ssa_test.py --skip-error tests/*/*
+    - python3 bin/ssa_validate.py --skip-error detections/*/*
     - python3 bin/testing_coverage.py --type SSA --min-coverage 1.0
